app/requirements.txt
fastapi==0.103.2
gunicorn==21.2.0
httpx==0.25.0
joblib==1.3.1
numpy==1.25.1
pydantic==2.4.2
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.0
timm==0.9.5
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
